+++ topiary/__init__.py
@@ -62,7 +62,7 @@ from .io_lens import detect_lens_version, read_lens
 from .result import TopiaryResult, concat
 from .wide import detect_form, from_wide, to_wide
 
-__version__ = "5.14.1"
+__version__ = "5.15.0"
 
 __all__ = [
     "TopiaryPredictor",